+++ src/lib/repository/find/helpers/validate-primary-columns.ts
@@ -1,11 +1,11 @@
-import type { SingleFindConditions } from "@techmmunity/symbiosis";
-import { SymbiosisError } from "@techmmunity/symbiosis";
+import type { SingleFindConditions } from "@thothom/core";
+import { ThothError } from "@thothom/core";
 import { cleanObj, getTypeof, isEmptyObject } from "@techmmunity/utils";
 
 import type { ExtraMetadata } from "../../../types/extra-metadata";
-import type { ColumnMetadata } from "@techmmunity/symbiosis/lib/entity-manager/types/column-metadata";
-import type { EntityMetadata } from "@techmmunity/symbiosis/lib/entity-manager/types/entity-metadata";
-import type { DatabaseEntity } from "@techmmunity/symbiosis/lib/types/database-entity";
+import type { ColumnMetadata } from "@thothom/core/lib/entity-manager/types/column-metadata";
+import type { EntityMetadata } from "@thothom/core/lib/entity-manager/types/entity-metadata";
+import type { DatabaseEntity } from "@thothom/core/lib/types/database-entity";
 
 interface ValidatePrimaryColumns {
 	startFrom: Partial<DatabaseEntity>;
@@ -14,9 +14,9 @@ interface ValidatePrimaryColumns {
 }
 
 const throwDefaultError = (extraDetails?: Array<any>) => {
-	throw new SymbiosisError({
+	throw new ThothError({
 		code: "INVALID_PARAM",
-		origin: "SYMBIOSIS",
+		origin: "THOTHOM",
 		message: "Invalid Params",
 		details: [
 			"To use the startFrom option, you must specify ALL AND ONLY the primary columns of the last value of previous operation.",