+++ src/lib/types/context.ts
@@ -1,30 +1,30 @@
 import type { DynamoDBClient } from "@aws-sdk/client-dynamodb";
-import type { EntityManager, Logger } from "@techmmunity/symbiosis";
-import type { AfterDeleteParams } from "@techmmunity/symbiosis/lib/repository/methods/delete/after";
+import type { EntityManager, Logger } from "@thothom/core";
+import type { AfterDeleteParams } from "@thothom/core/lib/repository/methods/delete/after";
 import type {
 	BeforeDeleteInput,
 	BeforeDeleteOutput,
-} from "@techmmunity/symbiosis/lib/repository/methods/delete/before";
-import type { AfterFindOneParams } from "@techmmunity/symbiosis/lib/repository/methods/find-one/after";
+} from "@thothom/core/lib/repository/methods/delete/before";
+import type { AfterFindOneParams } from "@thothom/core/lib/repository/methods/find-one/after";
 import type {
 	BeforeFindOneInput,
 	BeforeFindOneOutput,
-} from "@techmmunity/symbiosis/lib/repository/methods/find-one/before";
-import type { AfterFindParams } from "@techmmunity/symbiosis/lib/repository/methods/find/after";
+} from "@thothom/core/lib/repository/methods/find-one/before";
+import type { AfterFindParams } from "@thothom/core/lib/repository/methods/find/after";
 import type {
 	BeforeFindInput,
 	BeforeFindOutput,
-} from "@techmmunity/symbiosis/lib/repository/methods/find/before";
-import type { AfterSaveParams } from "@techmmunity/symbiosis/lib/repository/methods/save/after";
+} from "@thothom/core/lib/repository/methods/find/before";
+import type { AfterSaveParams } from "@thothom/core/lib/repository/methods/save/after";
 import type {
 	BeforeSaveInput,
 	BeforeSaveOutput,
-} from "@techmmunity/symbiosis/lib/repository/methods/save/before";
-import type { AfterUpsertParams } from "@techmmunity/symbiosis/lib/repository/methods/upsert/after";
+} from "@thothom/core/lib/repository/methods/save/before";
+import type { AfterUpsertParams } from "@thothom/core/lib/repository/methods/upsert/after";
 import type {
 	BeforeUpsertInput,
 	BeforeUpsertOutput,
-} from "@techmmunity/symbiosis/lib/repository/methods/upsert/before";
+} from "@thothom/core/lib/repository/methods/upsert/before";
 
 import type { ExtraMetadata } from "./extra-metadata";