+++ src/lib/types/methods-outputs/delete.ts
@@ -1,3 +1,3 @@
-import type { BaseDeleteOutput } from "@techmmunity/symbiosis";
+import type { BaseDeleteOutput } from "@thothom/core";
 
 export type DeleteOutput = BaseDeleteOutput;