+++ src/lib/types/methods-outputs/find.ts
@@ -1,4 +1,4 @@
-import type { BaseFindOutput, ClassType } from "@techmmunity/symbiosis";
+import type { BaseFindOutput, ClassType } from "@thothom/core";
 
 export interface FindOutput<T> extends BaseFindOutput<T> {
 	cursor?: ClassType<T>;