+++ src/lib/types/methods-outputs/recover.ts
@@ -1,3 +1,3 @@
-import type { BaseRecoverOutput } from "@techmmunity/symbiosis";
+import type { BaseRecoverOutput } from "@thothom/core";
 
 export type RecoverOutput = BaseRecoverOutput;