+++ src/lib/types/methods-outputs/soft-delete.ts
@@ -1,3 +1,3 @@
-import type { BaseSoftDeleteOutput } from "@techmmunity/symbiosis";
+import type { BaseSoftDeleteOutput } from "@thothom/core";
 
 export type SoftDeleteOutput = BaseSoftDeleteOutput;