+++ src/lib/types/methods-outputs/update.ts
@@ -1,3 +1,3 @@
-import type { BaseUpdateOutput } from "@techmmunity/symbiosis";
+import type { BaseUpdateOutput } from "@thothom/core";
 
 export type UpdateOutput<T> = BaseUpdateOutput<T>;