+++ src/lib/utils/get-find-command/index.ts
@@ -1,9 +1,9 @@
 import { QueryCommand, ScanCommand } from "@aws-sdk/client-dynamodb";
-import type { EntityManager, FindConditions } from "@techmmunity/symbiosis";
+import type { EntityManager, FindConditions } from "@thothom/core";
 import { getTypeof, isEmptyArray, isEmptyObject } from "@techmmunity/utils";
 
 import type { ExtraMetadata } from "../../types/extra-metadata";
-import type { DatabaseEntity } from "@techmmunity/symbiosis/lib/types/database-entity";
+import type { DatabaseEntity } from "@thothom/core/lib/types/database-entity";
 
 interface GetWhereCommandParams<Entity> {
 	where?: FindConditions<DatabaseEntity>;