+++ src/lib/utils/handle-database-error/index.ts
@@ -1,4 +1,4 @@
-import { SymbiosisError } from "@techmmunity/symbiosis";
+import { ThothError } from "@thothom/core";
 
 import type { Context } from "../../types/context";
 
@@ -6,17 +6,17 @@ export const handleDatabaseError = (context: Context<any>, err: any) => {
 	const message = err.message;
 
 	switch (true) {
-		case err instanceof SymbiosisError:
+		case err instanceof ThothError:
 			return err;
 		case message === "Requested resource not found":
-			return new SymbiosisError({
+			return new ThothError({
 				code: "OPERATION_FAILED",
 				origin: "DATABASE",
 				message: "Table not found",
 				details: [`Table "${context.tableName}" not found.`],
 			});
 		default:
-			return new SymbiosisError({
+			return new ThothError({
 				code: "UNKNOWN",
 				origin: "DATABASE",
 				message: "Unknown error",