+++ CHANGELOG.md
@@ -15,7 +15,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-## [0.0.18] - 2022-02-14
+### Dependencies
+
+## [0.0.1] - 2022-05-30
+
+### Added
+
+### Changed
+
+- Package name to `ThothOM`, `@thothom/dynamodb`
+
+### Fixed
+
+### Removed
+
+### Dependencies
+
+## [@techmmunity/symbiosis-dynamodb@0.0.18] - 2022-02-14
 
 ### Added
 
@@ -27,7 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-## [0.0.17] - 2021-12-05
+## [@techmmunity/symbiosis-dynamodb@0.0.17] - 2021-12-05
 
 ### Added
 
@@ -39,7 +55,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-## [0.0.16] - 2021-11-12
+## [@techmmunity/symbiosis-dynamodb@0.0.16] - 2021-11-12
 
 ### Added
 
@@ -52,7 +68,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-## [0.0.15] - 2021-11-04
+## [@techmmunity/symbiosis-dynamodb@0.0.15] - 2021-11-04
 
 ### Added